models/tongue-modeling/requirements.txt
absl-py==2.1.0
blinker==1.8.2
boto3==1.35.35
botocore==1.35.35
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dill==0.3.9
filelock==3.16.1
filetype==1.2.0
fire==0.7.0
Flask==3.0.3
fonttools==4.54.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.66.2
h11==0.14.0
huggingface-hub==0.24.7
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==3.0.0
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==10.4.0
protobuf==5.28.2
psutil==6.0.0
py-cpuinfo==9.0.0
pybboxes==0.1.6
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.47
s3transfer==0.10.2
safetensors==0.4.5
sahi==0.11.18
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentence-transformers==3.1.1
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tokenizers==0.20.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.7
ultralytics-thop==2.0.9
urllib3==2.2.3
uvicorn==0.31.0
Werkzeug==3.0.4
yolov5==7.0.13
